Fix deprecation of numpy.float and numpy.int (#734)
* Try new numpy package

* Revert "Try new numpy package"

This reverts commit 630d437915bc7ea49a29bffaa1f574b6319e9c5d.

* Remove requirement on numpy in setup.py

* Bump numpy version in requirements/requirements_install.txt

* Update np.int with np.int32

* Update np.float references of dpf-core source code

* Update np.float references in tests

* Add --find-links to docs, examples, gate, post CI templates that it can install wheels from standalone branch

* Add gate in .github folder to test it

* Remove gate from .github

* Update unit test in test_geometry

* Update docker tests run with extra-pip-args

* Add dpf_wheels_directory to CI (tests, for retro)

* Add ansys-dpf-gate wheel

* update dtype in docstring

* to revert: test

* Update test

* Add check

* add test

* Add test

* Move check step

* Update dpf_wheels_directory

* add check

* update ci

* Move check

* Update docker step

* test adding ROOT_SPACE

* Update CI

* Try source dir

* remove failing check

* Removing failing on docstring (different on windows and on linux)

* Docker step updated

* Redo Docker_test

* Update test_geometry.py

* Fix expected raised error type in test_geometry.py

* Remove numpy pin in pyproject.toml

* Fix expected raised error type in test_geometry.py

* Remove test gate package since it has been released

---------

+++ src/ansys/dpf/core/nodes.py
@@ -322,7 +322,7 @@ class Nodes:
             raise ValueError('Input scope location must be "Nodal"')
         arr = np.array(list(map(self.mapping_id_to_index.get, external_scope.ids)))
         mask = arr != None
-        ind = arr[mask].astype(np.int)
+        ind = arr[mask].astype(np.int32)
         return ind, mask
 
     def add_node(self, id, coordinates):
